ipython notebook for AutoEncoder
AutoEncoder is useful for learning feature without labels.
The process is breakdown the image itself by leaning the main features, and then reconstruct itself using these learned features.
There is only 1 hidden layer between input and output layers, which are of the same size. while the hidden layer has smaller size. The size of hidden layer is the number of "classes" for the images, or the "filters" in the NN that will be discussed later.